+++ app/lib/core/theme.dart
@@ -5,6 +5,7 @@ import 'package:dynamic_color/dynamic_color.dart';
 import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:provider/provider.dart';
+import 'package:trale/core/contrast.dart';
 
 import 'package:trale/core/traleNotifier.dart';
 import 'package:trale/l10n-gen/app_localizations.dart';
@@ -62,6 +63,7 @@ class TraleTheme {
     required this.seedColor,
     required this.brightness,
     this.isAmoled=false,
+    this.contrast=0.0,
   });
 
   /// copyWith constructor
@@ -69,11 +71,13 @@ class TraleTheme {
     Brightness? brightness,
     Color? seedColor,
     bool? isAmoled,
+    double? contrast,
   }) {
     return TraleTheme(
       brightness: brightness ?? this.brightness,
       seedColor: seedColor ?? this.seedColor,
       isAmoled: isAmoled ?? this.isAmoled,
+      contrast: contrast ?? this.contrast,
     );
   }
 
@@ -100,6 +104,8 @@ class TraleTheme {
   final double padding = 16;
   /// if true make background true black
   late bool isAmoled;
+  /// contrast level
+  late double contrast;
   /// Get border radius
   double get borderRadius => 16;
 
@@ -122,7 +128,9 @@ class TraleTheme {
 
   /// get elevated shade of clr
   Color colorOfElevation(double elevation, Color clr) => Color.alphaBlend(
-    getFontColor(clr).withOpacity(overlayOpacity(elevation)), clr,
+    getFontColor(clr).withValues(
+        alpha: overlayOpacity(elevation)
+    ), clr,
   );
   /// 24 elevation shade of bg
   Color get bgShade1 => bgElevated(24);
@@ -166,6 +174,7 @@ class TraleTheme {
     ColorScheme colorScheme = ColorScheme.fromSeed(
       seedColor: seedColor,
       brightness: brightness,
+      contrastLevel: contrast,
       dynamicSchemeVariant: isGrey
         ? DynamicSchemeVariant.fidelity
         : DynamicSchemeVariant.tonalSpot,
@@ -189,6 +198,10 @@ class TraleTheme {
       year2023: false,
     );
 
+    const CardThemeData cardTheme = CardThemeData(
+      shadowColor: Colors.transparent,
+    );
+
     const ProgressIndicatorThemeData progressIndicatorTheme =
       ProgressIndicatorThemeData(year2023: false);
 
@@ -200,6 +213,7 @@ class TraleTheme {
     ).copyWith(
       listTileTheme: listTileThemeData,
       sliderTheme: sliderTheme,
+      cardTheme: cardTheme,
       progressIndicatorTheme: progressIndicatorTheme,
       pageTransitionsTheme: const PageTransitionsTheme(
         builders: <TargetPlatform, PageTransitionsBuilder>{
@@ -257,14 +271,22 @@ extension TraleCustomThemeExtension on TraleCustomTheme {
     TraleCustomTheme.plum: const Color(0xff8e4585),
   }[this]!;
 
+  /// get contrast level
+  double contrast(BuildContext context) =>
+    Provider.of<TraleNotifier>(context, listen: false).contrastLevel.contrast;
+
   /// get corresponding light theme
   TraleTheme light(BuildContext context) => TraleTheme(
-    seedColor: seedColor(context), brightness: Brightness.light,
+    seedColor: seedColor(context),
+    brightness: Brightness.light,
+    contrast: contrast(context),
   );
 
   /// get corresponding light theme
   TraleTheme dark(BuildContext context) => TraleTheme(
-    seedColor: seedColor(context), brightness: Brightness.dark,
+    seedColor: seedColor(context),
+    brightness: Brightness.dark,
+    contrast: contrast(context),
   );
 
   /// get amoled dark theme
@@ -370,4 +392,17 @@ class TraleThemeExtension extends ThemeExtension<TraleThemeExtension> {
       padding: lerpDouble(padding, other.padding, t)
     );
   }
+}
+
+/// extension of theme
+extension ColorTextThemeExtension on TextStyle {
+  TextStyle onSecondaryContainer(BuildContext context) => copyWith(
+    color: Theme.of(context).colorScheme.onSecondaryContainer,
+  );
+  TextStyle onSurface(BuildContext context) => copyWith(
+    color: Theme.of(context).colorScheme.onSurface,
+  );
+  TextStyle onSurfaceVariant(BuildContext context) => copyWith(
+    color: Theme.of(context).colorScheme.onSurfaceVariant,
+  );
 }